chore: scene tweaks for car physics test + AtlasQuadTreePoint cast warning fix
CollisionTest.Scene에 ACarPawn 배치/속성 조정 (PhysX compound 동작 검증용).
AtlasQuadTreePoint.cpp의 float→int32 인자 캐스트 명시화로 C4244 경고 제거.

+++ KraftonEngine/Source/Engine/Render/Shadow/AtlasQuadTreePoint.cpp
@@ -37,7 +37,7 @@ TArray<FAtlasRegion> FAtlasQuadTreePoint::CommitBatch() {
 		auto desired_res = std::min(Batch[OrigIdx].Resolution, AtlasSize);
 		desired_res = static_cast<float>(RoundToNearestPowerOfTwo(static_cast<uint32>(desired_res)));
 		desired_res = desired_res > MinShadowMapResolution ? desired_res : MinShadowMapResolution;
-		FAtlasRegion AtlasRegion = AllocateNode(0, desired_res, Batch[OrigIdx].LightIdx, Batch[OrigIdx].Light.CubeMapOrientation);
+		FAtlasRegion AtlasRegion = AllocateNode(0, (int32)desired_res, Batch[OrigIdx].LightIdx, Batch[OrigIdx].Light.CubeMapOrientation);
 		Results[OrigIdx] = AtlasRegion;
 	}
